+++ util/Coordinates.py
@@ -1,0 +1,8 @@
+
+
+
+class WorldPoint:
+    def __init__(self, x: int=0, y: int=0, z: int=0) -> None:
+        self.x = x
+        self.y = y
+        self.z = z
